refactor(Step5) : domain 객체 위치 조정
패키지 조정 :
  - RacingGame 클래스 domain으로 이동.
