+++ src/test/java/racingGame/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racingGame.logic;
+package racingGame.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
